refactor: createLesson route
